Update patch set 1
Patch Set 1:

> > > > > P.S. you might even remove it completely in mr5.5, AFAIK it
 > > has
 > > > > > been removed in mr6.2+ already by Guillem.
 > > > >
 > > > > > P.S. you might even remove it completely in mr5.5, AFAIK it
 > > has
 > > > > > been removed in mr6.2+ already by Guillem.
 > > > >
 > > > > Yeah was going to say that, we do not support elasticsearch
 > any
 > > > > more.
 > > >
 > > > In which version was it removed?
 > >
 > > The support in templates got removed in commit 32b67fe97660f09c4124b624457ebc0a17a4e5cf
 > > which was mr6.2.
 > > I might have missed other instances in other places, perhaps
 > > filename-based references. :/
 > 
 > So it's still supported in 5.5? Getting confused here

I think my usage of support was probably confusing. I think what Alex meant (and he'll correct me if I'm wrong) is that we never really supported elastichsearch, we had some code in some releases though, but I'm not sure anyone really tested the thing as a whole or used it in production. That's why I think Alex is saying that we could even remove it from mr5.5?

Patch-set: 1