MT#55524 Update lintian overrides for new lintian behavior
lintian v2.115.3 - as currently present in Debian/booksworm - doesn't
support the existing syntax for the non-standard-toplevel-dir override,
and therefore fails with:

| E: ngcp-cleanup-tools-ce: non-standard-toplevel-dir [ngcp-data/]
| E: ngcp-cleanup-tools-pro: non-standard-toplevel-dir [ngcp-data/]

It instead expects the filename(s) to be listed under "[...]".
By using "*...*" instead, we can support the lintian versions as present
in bookworm and bullseye both at the same time.

Change-Id: Ie972616187981f29d4feb5fc3e7c94ca4d9d9281